feat(frontend): add reusable recipe list component
